Minor edits
