+++ src/main/java/busynessmanager/BusynessManager.java
@@ -15,6 +15,19 @@ public class BusynessManager {
         System.out.println("Hello from\n" + logo);
         System.out.println("What is your name?");
 
+        // test code
+        Product product1 = new Product("beans", 1, 0.6);
+        Product product2 = new Product("donkey", 6, 200);
+
+        InventoryManager im = new InventoryManager();
+        im.addProduct(product1.getName(), product1.getQuantity(), product1.getPrice());
+        im.addProduct(product2.getName(), product2.getQuantity(), product2.getPrice());
+
+
+        im.printProducts();
+
+
+
         Scanner in = new Scanner(System.in);
         System.out.println("Hello " + in.nextLine());
     }
